requirements.txt
# Compliance Copilot - Day 2 Pipeline Dependencies
# Core PDF processing
PyPDF2>=3.0.0

# AI and ML capabilities
google-generativeai>=0.3.0

# Data processing and analysis
pandas>=1.5.0

# Web scraping and HTTP requests (for future use)
requests>=2.28.0
beautifulsoup4>=4.11.0
lxml>=4.9.0

# PDF generation and reporting
reportlab>=3.6.0

# Standard library dependencies (included with Python 3.9+)
# json, logging, pathlib, datetime, re, argparse, collections
# For loading environment variables from .env file
python-dotenv>=0.21.0